requirements.txt
opencv-python
numpy
python-chess
cairosvg
